verifying e2e pipeline works
Automatically inferring go package name from service name (convention
over configuration). As a result renaming the package to the service,
which may be more flexible in allowing for multiple generated services
to live side-by-side without wrapping packages in an outer layer.

Choosing "Read" instead of "Get", so that the acronym "CRUDL" can be
used to map to each of the RPCs.